feat: generate shared on demand

+++ packages/module-federation-metro/src/plugin.ts
@@ -209,20 +209,21 @@ function createSharedRegistryVirtualModule(
   return sharedRegistryPath;
 }
 
-function createSharedVirtualModules(
-  options: ModuleFederationConfigNormalized,
-  vmDirPath: string
-) {
-  const sharedModulesPaths: Record<string, string> = {};
-  Object.keys(options.shared).forEach((name) => {
-    const sharedModule = getSharedModule(name);
-    const sharedFilePath = path.join(vmDirPath, "shared", `${name}.js`);
-    // needed for deep imports
+function createSharedModule(sharedName: string, outputDir: string) {
+  const sharedFilePath = getSharedPath(sharedName, outputDir);
+  // we need to create the shared module if it doesn't exist
+  if (!fs.existsSync(sharedFilePath)) {
+    const sharedModule = getSharedModule(sharedName);
     fs.mkdirSync(path.dirname(sharedFilePath), { recursive: true });
     fs.writeFileSync(sharedFilePath, sharedModule, "utf-8");
-    sharedModulesPaths[name] = sharedFilePath;
-  });
-  return sharedModulesPaths;
+  }
+  return sharedFilePath;
+}
+
+function getSharedPath(name: string, dir: string) {
+  const sharedName = name.replaceAll("/", "_");
+  const sharedDir = path.join(dir, "shared");
+  return path.join(sharedDir, `${sharedName}.js`);
 }
 
 function replaceModule(from: RegExp, to: string) {
@@ -289,8 +290,6 @@ function withModuleFederation(
     options,
     mfMetroPath
   );
-
-  const sharedModulesPaths = createSharedVirtualModules(options, mfMetroPath);
 
   const initHostPath = isHost
     ? createInitHostVirtualModule(options, mfMetroPath)
@@ -389,17 +388,11 @@ function withModuleFederation(
           const sharedModule = options.shared[moduleName];
           // import: false means that the module is marked as external
           if (sharedModule && sharedModule.import === false) {
-            const sharedPath = sharedModulesPaths[moduleName];
+            const sharedPath = getSharedPath(moduleName, mfMetroPath);
             return { type: "sourceFile", filePath: sharedPath };
           } else {
             return context.resolveRequest(context, moduleName, platform);
           }
-        }
-
-        // shared modules
-        if (Object.keys(options.shared).includes(moduleName)) {
-          const sharedPath = sharedModulesPaths[moduleName];
-          return { type: "sourceFile", filePath: sharedPath };
         }
 
         // replace getDevServer module in remote with our own implementation
@@ -409,6 +402,21 @@ function withModuleFederation(
             /react-native\/Libraries\/Core\/Devtools\/getDevServer\.js$/;
           const to = path.resolve(__dirname, "../getDevServer.js");
           return replaceModule(from, to)(res);
+        }
+
+        // shared module handling
+        for (const sharedName of Object.keys(options.shared)) {
+          const importName = options.shared[sharedName].import || sharedName;
+          // module import
+          if (moduleName === importName) {
+            const sharedPath = createSharedModule(moduleName, mfMetroPath);
+            return { type: "sourceFile", filePath: sharedPath };
+          }
+          // module deep import
+          if (importName.endsWith("/") && moduleName.startsWith(importName)) {
+            const sharedPath = createSharedModule(moduleName, mfMetroPath);
+            return { type: "sourceFile", filePath: sharedPath };
+          }
         }
 
         return context.resolveRequest(context, moduleName, platform);
